Another Zomboid server Rcon script
These are Bash scripts for automating a Zomboid server on Linux. It includes Rcon support for restarting, shutting down, and backing up the server. start-server.sh is to replace the default start-server.sh provided. It includes an additional line which creates a backup of your server at startup.

Please refer to this page: https://shockbyte.com/billing/knowledgebase/486/How-to-Setup-RCON-for-Your-Project-Zomboid-Server.html
For more information on setting up RCON on your server.